+++ packages/ui/src/menu/link-item.tsx
@@ -17,6 +17,7 @@ const LinkItem = forwardRef< Element, LinkItemProps >( function MenuLinkItem(
 		children,
 		className,
 		openInNewTab = false,
+		target,
 		prefix,
 		shortcut,
 		suffix,
@@ -29,7 +30,8 @@ const LinkItem = forwardRef< Element, LinkItemProps >( function MenuLinkItem(
 	},
 	ref
 ) {
-	const externalLinkIndicator = openInNewTab ? (
+	const shouldShowNewTabIndicator = openInNewTab || target === '_blank';
+	const externalLinkIndicator = shouldShowNewTabIndicator ? (
 		<span
 			className={ styles[ 'external-link-indicator' ] }
 			role="img"
@@ -39,15 +41,19 @@ const LinkItem = forwardRef< Element, LinkItemProps >( function MenuLinkItem(
 			}
 		/>
 	) : null;
-	const { contentContextValue, itemAriaProps, shortcutDescriptionId } =
-		useItemContent( children, {
-			'aria-describedby': ariaDescribedBy,
-			'aria-keyshortcuts': ariaKeyShortcuts,
-			'aria-label': ariaLabel,
-			'aria-labelledby': ariaLabelledBy,
-			labelTrailing: externalLinkIndicator,
-			shortcut,
-		} );
+	const {
+		contentChildren,
+		contentContextValue,
+		itemAriaProps,
+		shortcutDescriptionId,
+	} = useItemContent( children, {
+		'aria-describedby': ariaDescribedBy,
+		'aria-keyshortcuts': ariaKeyShortcuts,
+		'aria-label': ariaLabel,
+		'aria-labelledby': ariaLabelledBy,
+		labelTrailing: externalLinkIndicator,
+		shortcut,
+	} );
 
 	return (
 		<_Menu.LinkItem
@@ -55,7 +61,7 @@ const LinkItem = forwardRef< Element, LinkItemProps >( function MenuLinkItem(
 			{ ...props }
 			{ ...itemAriaProps }
 			rel={ rel }
-			target={ openInNewTab ? '_blank' : undefined }
+			target={ target ?? ( openInNewTab ? '_blank' : undefined ) }
 			className={ clsx(
 				defenseStyles.a,
 				resetStyles[ 'box-sizing' ],
@@ -70,7 +76,7 @@ const LinkItem = forwardRef< Element, LinkItemProps >( function MenuLinkItem(
 					shortcutDescriptionId={ shortcutDescriptionId }
 					suffix={ suffix }
 				>
-					{ children }
+					{ contentChildren }
 				</ItemContent>
 			</MenuItemContentContext.Provider>
 		</_Menu.LinkItem>